array by replacing arr[i] with (arr[i-1]+1) % M
Reconstruct the array by replacing arr[i] with (arr[i-1]+1) % M
Given an array of N elements and an integer M. Now, the array is modified by replacing some of the array elements with -1. The task is to print the original array.

The elements in the orginal array are related as, for every index i, a[i] = (a[i-1]+1)% M.

It is guaranteed that there is one non zero value in the array.